move App.js and relative components to a folder
