+++ src/components/app/App.css
@@ -27,14 +27,10 @@ html body{
  :global .true{
     transform:translate(0rem ,24rem);
  }
-
-
 ul{
   list-style:none;
-  
 }
-
-
+
