+++ src/components/app/App.js
@@ -1,14 +1,14 @@
 import './App.css';
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import SplashPage from './splash-page/SplashPage';
-import Nav from './nav/Nav';
-import Expertise from './expertise/Expertise';
-import Background from './background/Background';
-import Contact from './contact/Contact';
-import Footer from './footer/Footer';
-import Services from './services/Services';
-import MobileNavHook from './nav/MobileNavHook';
+import SplashPage from '../splash-page/SplashPage';
+import Nav from '../nav/Nav';
+import Expertise from '../expertise/Expertise';
+import Background from '../background/Background';
+import Contact from '../contact/Contact';
+import Footer from '../footer/Footer';
+import Services from '../services/Services';
+import MobileNavHook from '../nav/MobileNavHook';
 
 export default function App() {
   return (
