+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import WebFont from 'webfontloader';
-import App from './components/App';
+import App from './components/app/App';
 
 WebFont.load({
   google: {
